# Various (minor) tweaks in regards to allocations and performance