First off, thank you for the work put into this.

I wanted to drop this library into my code to try it out (i write a performance-pressured code generator) but unfortunately it appears loops are not supported as part of 2.0 (I saw the PR for it, though, so that's nice). Anyways.

I noticed a few minor issues while looking through the code.
* `Expression.ToParamsCode<T>` uses string concatenation instead of a `StringBuilder`. This is probably not very important, but it's nicer now.
* `Expression.Block` is missing an overload accepting a `List<Expression>` for its body. Not a huge deal, but annoying, since `T[] ToArray()` generally implies an array copy (And this also contrasts with other static methods - note I did *not* check the overloads for `Linq.Expression`)
* `Tools.AsArray<T>(this IEnumerable<T> source)` uses a temporary list to load the enumerable and then turn it into a `T[]` (if source is not `T[]`). This is not pretty. because again, there is an implicit copy during `ToArray()`, on top of `List<T>` performing any possible number of resizes and copies when loading elements from the `IEnumerable<T>` it is provided.
This is the only one I actually wrote a benchmark for.

```
BenchmarkDotNet=v0.11.5, OS=Windows 10.0.18362
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=2.2.401
  [Host]     : .NET Core 2.2.6 (CoreCLR 4.6.27817.03, CoreFX 4.6.27818.02), 64bit RyuJIT
  DefaultJob : .NET Core 2.2.6 (CoreCLR 4.6.27817.03, CoreFX 4.6.27818.02), 64bit RyuJIT


|            Method | rangeMax |         Mean |         Error |        StdDev |    Gen 0 |    Gen 1 |    Gen 2 | Allocated |
|------------------ |--------- |-------------:|--------------:|--------------:|---------:|---------:|---------:|----------:|
| EnumerableToArray |      100 |     101.7 ns |      2.072 ns |      1.938 ns |   0.1347 |        - |        - |     424 B |
| EnumerableAsArray |      100 |     967.0 ns |     18.913 ns |     21.022 ns |   0.5245 |        - |        - |    1656 B |
| EnumerableToArray |     1000 |     889.0 ns |      4.192 ns |      3.921 ns |   1.2779 |        - |        - |    4024 B |
| EnumerableAsArray |     1000 |   7,154.9 ns |     21.956 ns |     18.334 ns |   3.9673 |        - |        - |   12496 B |
| EnumerableToArray |    10000 |   7,400.4 ns |    142.944 ns |    170.164 ns |  12.6495 |        - |        - |   40024 B |
| EnumerableAsArray |    10000 |  71,768.9 ns |    903.127 ns |    844.785 ns |  53.9551 |        - |        - |  171472 B |
| EnumerableToArray |   100000 |  75,608.7 ns |  1,241.826 ns |  1,100.846 ns | 124.8779 | 124.8779 | 124.8779 |  400024 B |
| EnumerableAsArray |   100000 | 810,969.4 ns | 13,054.779 ns | 11,572.722 ns | 399.4141 | 399.4141 | 399.4141 | 1449048 B |
```

And the benchmark itself: https://gist.github.com/Warpten/982949dfc362cff78b35c1b2c9668298

Now, I'm not sure whether or not there is an actual reason for this specific implementation, but cutting the temporary makes the code noticeably faster (although it is by no means a hot path for regular usage), and less allocatey: we get exactly what is expected from `ToArray<T>()` : 24 bytes for the overhead of `T[]`, and 4 bytes per element.

* And finally, I added some overloads to `GetFirst<T>`, to cut boxing. I also assume `GetFirst<T>()` is used in place of LINQ's `FirstOrDefault()` to avoid the `IPartition<T>` check. For now, this only affects one call site: `CodePrinter.ToCode`.

Now, I'm not sure what the best way to contribute; these are not huge issues, so I'm not sure splitting them into different pull requests would be welcomed. Additionally, should I add benchmarks for any of these? Please let me know.