@dadhi 2019-09-15
Hi, thanks for PR. It seems fine!

Just a note regarding `CodePrinter.ToCode`,
it supposed to be called for debugging purposes to drop result expression into the test or something. That is why I did not bothered to optimize it.